feat: Conditionally display "Douane Users Table" button
This commit implements conditional rendering of the "Tableau des utilisateurs Douane" button in Chart.tsx and Chart4All.tsx. The button is now only displayed when the selected service is 'Douane'. This change ensures that the button is only visible when relevant to the displayed data. Also removes the now obsolete HistoriqueDesModifications_6.md file.